bfdd: hide echo-mode command
Echo-mode implementation is currently broken. Instead of sending packets
to it's own address, bfdd is sending echo packets to the peer's address.
It may seem to work when testing between two FRR instances, because FRR
loops back such packets, but no other implementation is supposed to do
that. So it's better to hide echo-mode command until this mode is fixed
and throw a warning when user tries to enable this mode.

Signed-off-by: Igor Ryzhov <iryzhov@nfware.com>